python/observability-online-eval/openai/computer-use/pyproject.toml
[project]
name = "computer-use"
version = "0.1.0"
description = "Langgraph CUA logging via Maxim"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph-cua==0.0.1",
    "langchain==0.3.26",
    "openai==1.91.0",
    "scrapybara==2.5.3",
    "maxim-py==3.9.6",
    "python-dotenv>=1.1.1",
]
